requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
discord.py
aiohttp
